requirements.txt
langchain
openai
google-generativeai
langgraph
pypdf
python-pptx
pandas
